feat: initialize plugin
